less change on polling
