+++ jvm/core/main/scala/com/webank/eggroll/core/constant/ConfKeys.scala
@@ -183,9 +183,11 @@ object RollSiteConfKeys {
   val EGGROLL_ROLLSITE_POLLING_CLIENT_ENABLED = ErConfKey("eggroll.rollsite.polling.client.enabled", false)
   val EGGROLL_ROLLSITE_POLLING_CONCURRENCY = ErConfKey("eggroll.rollsite.polling.concurrency", 50)
   val EGGROLL_ROLLSITE_ROUTE_TABLE_KEY = ErConfKey("eggroll.rollsite.route.table.key", "123")
-  val EGGROLL_ROLLSITE_POLLING_AHTHENTICATION_ENABLE = ErConfKey("eggroll.rollsite.polling.authentication.enable", false)
+  val EGGROLL_ROLLSITE_POLLING_AHTHENTICATION_ENABLED = ErConfKey("eggroll.rollsite.polling.authentication.enabled", false)
+  val EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_USE_CONFIGURE = ErConfKey("eggroll.rollsite.polling.authentication.use.configure", false)
   val EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_URI = ErConfKey("eggroll.rollsite.polling.authentication.uri", "/cloud-manager/api/site/rollsite/checkPartyId")
   val EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_SECRECT_INFO_GENERATOR = ErConfKey("eggroll.rollsite.polling.authentication.secret.info.generator", "com.webank.ai.fate.cloud.sdk.sdk.Fatecloud#getSecretInfo")
+  val EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_SIGNATER_GENERATOR = ErConfKey("eggroll.rollsite.polling.authentication.signature.generator", "com.webank.ai.fate.cloud.sdk.sdk.Fatecloud#generateSignature")
   val EGGROLL_ROLLSITE_POLLING_AUTHENTICATOR_INTERFACE = ErConfKey("eggroll.rollsite.polling.authenticator.interface", "com.webank.ai.fate.cloud.sdk.sdk.Fatecloud#checkPartyId")
   val EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_APPKEY = ErConfKey("eggroll.rollsite.polling.authentication.appKey", "")
   val EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_APPSERCRET = ErConfKey("eggroll.rollsite.polling.authentication.appSecret", "")
