+++ jvm/roll_site/main/scala/com/webank/eggroll/rollsite/Polling.scala
@@ -29,10 +29,11 @@ import com.webank.eggroll.core.meta.ErRollSiteHeader
 import com.webank.eggroll.core.meta.TransferModelPbMessageSerdes.ErRollSiteHeaderFromPbMessage
 import com.webank.eggroll.core.transfer.GrpcClientUtils
 import com.webank.eggroll.core.transfer.Transfer.RollSiteHeader
-import com.webank.eggroll.core.util.{ErrorUtils, Logging, ToStringUtils, AuthenticationUtils}
+import com.webank.eggroll.core.util.{AuthenticationUtils, ErrorUtils, Logging, ToStringUtils}
 import com.webank.eggroll.rollsite.PollingResults.errorPoison
 import io.grpc.ConnectivityState
 import io.grpc.stub.{ServerCallStreamObserver, StreamObserver}
+import javax.security.sasl.AuthenticationException
 import org.apache.commons.lang3.StringUtils
 import org.apache.commons.lang3.reflect.MethodUtils
 
@@ -62,6 +63,8 @@ object LongPollingClient extends Logging {
 
   private val pollingSemaphore = new Semaphore(RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_CONCURRENCY.get().toInt)
 
+  private var pollingTerminate: Boolean = false
+
   def acquireSemaphore(): Unit = {
     LongPollingClient.pollingSemaphore.acquire()
   }
@@ -80,69 +83,70 @@ class LongPollingClient extends Logging {
       var isSecure = Router.query("default").isSecure
       val caCrt = CoreConfKeys.CONFKEY_CORE_SECURITY_CLIENT_CA_CRT_PATH.get()
 
-      val pollingAuthenticationEnable = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AHTHENTICATION_ENABLE.get().toBoolean
+      val pollingAuthenticationEnable = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AHTHENTICATION_ENABLED.get().toBoolean
       if (pollingAuthenticationEnable) {
         //generate signature
-        val myPartyId = RollSiteConfKeys.EGGROLL_ROLLSITE_PARTY_ID.get().toInt
-        val secretInfoUrl = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_SECRET_INFO_URL.get().toString
-        var appSecret = ""
-        var appKey = ""
-        var role = ""
-        val authInfoSecretGenerator = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_SECRECT_INFO_GENERATOR.get().toString
-
-        try {
-          val splitted = authInfoSecretGenerator.split("#")
-          val authenticator = Class.forName(splitted(0)).newInstance()
-          val args = secretInfoUrl + "," + myPartyId
-          val result = MethodUtils.invokeExactMethod(authenticator, splitted(1), args.split(","): _*).asInstanceOf[String]
-          if (result == null || result == "") {
-            throw new IllegalArgumentException(s"result of ${splitted(1)} is empty")
-          }
-
-          val secretInfo = new JSONObject(result.mkString)
-          logTrace(s"secretInfo from ${splitted(0)}:${secretInfo}")
-          if (secretInfo.getJSONObject("data")== null) {
-            logInfo(s"partyID:${myPartyId} not registered")
-          }
-
-          appSecret = secretInfo.getJSONObject("data").getString("appSecret")
-          appKey = secretInfo.getJSONObject("data").getString("appKey")
-          logTrace(s"role of ${myPartyId} is ${secretInfo.getJSONObject("data").getString("role")}")
-          role = if (secretInfo.getJSONObject("data").getString("role").toLowerCase() == "guest") "1" else "2"
-
-        } catch {
-          case e: NoSuchMethodException =>
-            throw new NoSuchMethodException(s"failed to execute reflection")
-          case t: Throwable =>
-            logTrace(s"failed to get secretInfo from ${authInfoSecretGenerator.split("#")(1)}, please check if service ${secretInfoUrl} is available. " +
-              "Now try to get secretInfo from eggroll.properties")
-
-            appKey = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_APPKEY.get().toString
-            appSecret = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_APPSERCRET.get().toString
-            role = if (RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_ROLE.get().toString.toLowerCase() == "guest") "1"
-            else if (RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_ROLE.get().toString.toLowerCase() == "host") "2"
-            else "3"
-
-            if (appKey == null || appSecret == null || role == "3") {
-              throw new IllegalArgumentException(s"failed to get appKey or appSecret or party role from eggroll.properties")
-            }
-        }
-
-        val time = String.valueOf(System.currentTimeMillis)
-        val uuid = UUID.randomUUID.toString
-        val nonce = uuid.replaceAll("-", "")
-        val httpURI = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_URI.get().toString
-        val body = ""
-        val signature = AuthenticationUtils.generateSignature(appSecret, String.valueOf(myPartyId), role,
-          appKey, time, nonce, httpURI, body)
-
-        val authInfo: JSONObject = new JSONObject
-        authInfo.put("signature", signature.toString)
-        authInfo.put("appKey", appKey.toString)
-        authInfo.put("timestamp", time.toString)
-        authInfo.put("nonce", nonce.toString)
-        authInfo.put("role", role.toString)
-        authInfo.put("httpUri", httpURI)
+//        val myPartyId = RollSiteConfKeys.EGGROLL_ROLLSITE_PARTY_ID.get().toInt
+//        val secretInfoUrl = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_SECRET_INFO_URL.get().toString
+//        var appSecret = ""
+//        var appKey = ""
+//        var role = ""
+//        val authInfoSecretGenerator = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_SECRECT_INFO_GENERATOR.get().toString
+//
+//        try {
+//          val splitted = authInfoSecretGenerator.split("#")
+//          val authenticator = Class.forName(splitted(0)).newInstance()
+//          val args = secretInfoUrl + "," + myPartyId
+//          val result = MethodUtils.invokeExactMethod(authenticator, splitted(1), args.split(","): _*).asInstanceOf[String]
+//          if (result == null || result == "") {
+//            throw new IllegalArgumentException(s"result of ${splitted(1)} is empty")
+//          }
+//
+//          val secretInfo = new JSONObject(result.mkString)
+//          logTrace(s"secretInfo from ${splitted(0)}:${secretInfo}")
+//          if (secretInfo.getJSONObject("data")== null) {
+//            logInfo(s"partyID:${myPartyId} not registered")
+//          }
+//
+//          appSecret = secretInfo.getJSONObject("data").getString("appSecret")
+//          appKey = secretInfo.getJSONObject("data").getString("appKey")
+//          logTrace(s"role of ${myPartyId} is ${secretInfo.getJSONObject("data").getString("role")}")
+//          role = if (secretInfo.getJSONObject("data").getString("role").toLowerCase() == "guest") "1" else "2"
+//
+//        } catch {
+//          case e: NoSuchMethodException =>
+//            throw new NoSuchMethodException(s"failed to execute reflection")
+//          case t: Throwable =>
+//            logTrace(s"failed to get secretInfo from ${authInfoSecretGenerator.split("#")(1)}, please check if service ${secretInfoUrl} is available. " +
+//              "Now try to get secretInfo from eggroll.properties")
+//
+//            appKey = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_APPKEY.get().toString
+//            appSecret = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_APPSERCRET.get().toString
+//            role = if (RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_ROLE.get().toString.toLowerCase() == "guest") "1"
+//            else if (RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_ROLE.get().toString.toLowerCase() == "host") "2"
+//            else "3"
+//
+//            if (appKey == null || appSecret == null || role == "3") {
+//              throw new IllegalArgumentException(s"failed to get appKey or appSecret or party role from eggroll.properties")
+//            }
+//        }
+//
+//        val time = String.valueOf(System.currentTimeMillis)
+//        val uuid = UUID.randomUUID.toString
+//        val nonce = uuid.replaceAll("-", "")
+//        val httpURI = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_URI.get().toString
+//        val body = ""
+//        val signature = AuthenticationUtils.generateSignature(appSecret, String.valueOf(myPartyId), role,
+//          appKey, time, nonce, httpURI, body)
+//
+//        val authInfo: JSONObject = new JSONObject
+//        authInfo.put("signature", signature.toString)
+//        authInfo.put("appKey", appKey.toString)
+//        authInfo.put("timestamp", time.toString)
+//        authInfo.put("nonce", nonce.toString)
+//        authInfo.put("role", role.toString)
+//        authInfo.put("httpUri", httpURI)
+        val pollingAuth = new PollingAuthentication
 
         LongPollingClient.defaultPollingReqMetadata = Proxy.Metadata.newBuilder()
           .setDst(
@@ -150,7 +154,7 @@ class LongPollingClient extends Logging {
               .setPartyId(LongPollingClient.defaultPollingReqMetadata.getDst.getPartyId))
           .setTask(Proxy.Task.newBuilder()
             .setModel(Proxy.Model.newBuilder()
-              .setDataKey(authInfo.toString())))
+              .setDataKey(pollingAuth.genSecretInfo())))
           .build()
 
         LongPollingClient.initPollingFrameBuilder = Proxy.PollingFrame.newBuilder().setMetadata(LongPollingClient.defaultPollingReqMetadata)
@@ -229,12 +233,13 @@ class LongPollingClient extends Logging {
       // TODO:0: configurable
     } catch {
       case t: Throwable =>
+        LongPollingClient.pollingTerminate = true
         logError("polling failed", t)
     }
   }
 
   def pollingForever(): Unit = {
-    while (true) {
+    while (LongPollingClient.pollingTerminate) {
       try {
         polling()
       } catch {
@@ -366,46 +371,63 @@ class DispatchPollingReqSO(eggSiteServicerPollingRespSO: ServerCallStreamObserve
     if (inited) return
     logTrace(s"DispatchPollingReqSO.ensureInited calling.")
 
-    val pollingAuthenticationEnable = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AHTHENTICATION_ENABLE.get().toBoolean
+//    val pollingAuthenticationEnable = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AHTHENTICATION_ENABLE.get().toBoolean
+//    if (pollingAuthenticationEnable) {
+//      val authUrl = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_URL.get().toString
+//      val authString = req.getMetadata.getTask.getModel.getDataKey
+//      if (authString == "" || authString == null) {
+//        throw new IllegalStateException(s"failed to get authentication info from header")
+//      }
+//      val authInfo = new JSONObject(authString)
+//
+//      val signature = authInfo.getString("signature")
+//      val appKey = authInfo.getString("appKey")
+//      val timestamp = authInfo.getString("timestamp")
+//      val nonce = authInfo.getString("nonce")
+//      val role = authInfo.getString("role")
+//      val httpUri = authInfo.getString("httpUri")
+//      val authPartyID = req.getMetadata.getDst.getPartyId
+//
+//      val heads = new util.HashMap[String, String]()
+//      heads.put("TIMESTAMP", timestamp)
+//      heads.put("PARTY_ID", authPartyID)
+//      heads.put("NONCE", nonce)
+//      heads.put("ROLE", role)
+//      heads.put("APP_KEY", appKey)
+//      heads.put("URI", httpUri)
+//      heads.put("SIGNATURE", signature)
+//      val body = ""
+//      logTrace(s"auth heads:${heads.toString}")
+//
+//      val authInterface = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATOR_INTERFACE.get().toString
+//      val splitted = authInterface.split("#")
+//      val authenticator = Class.forName(splitted(0)).newInstance()
+//      val result = MethodUtils.invokeExactMethod(authenticator, splitted(1), authUrl, heads, body).asInstanceOf[Boolean]
+//      if (result) {
+//        logTrace(s"polling authentication of party=${authPartyID} passed")
+//      } else {
+//        logError(s"polling authentication of party=${authPartyID} failed, please check polling client authentication info")
+//        throw new IllegalArgumentException(s"polling authentication of party=${authPartyID} failed, please check polling client authentication info")
+//      }
+//    } else {
+//      logDebug("polling authentication disable")
+//    }
+
+    val pollingAuthenticationEnable = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AHTHENTICATION_ENABLED.get().toBoolean
     if (pollingAuthenticationEnable) {
-      val authUrl = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATION_URL.get().toString
-      val authString = req.getMetadata.getTask.getModel.getDataKey
-      if (authString == "" || authString == null) {
-        throw new IllegalStateException(s"failed to get authentication info from header")
-      }
-      val authInfo = new JSONObject(authString)
-
-      val signature = authInfo.getString("signature")
-      val appKey = authInfo.getString("appKey")
-      val timestamp = authInfo.getString("timestamp")
-      val nonce = authInfo.getString("nonce")
-      val role = authInfo.getString("role")
-      val httpUri = authInfo.getString("httpUri")
-      val authPartyID = req.getMetadata.getDst.getPartyId
-
-      val heads = new util.HashMap[String, String]()
-      heads.put("TIMESTAMP", timestamp)
-      heads.put("PARTY_ID", authPartyID)
-      heads.put("NONCE", nonce)
-      heads.put("ROLE", role)
-      heads.put("APP_KEY", appKey)
-      heads.put("URI", httpUri)
-      heads.put("SIGNATURE", signature)
-      val body = ""
-      logTrace(s"auth heads:${heads.toString}")
-
-      val authInterface = RollSiteConfKeys.EGGROLL_ROLLSITE_POLLING_AUTHENTICATOR_INTERFACE.get().toString
-      val splitted = authInterface.split("#")
-      val authenticator = Class.forName(splitted(0)).newInstance()
-      val result = MethodUtils.invokeExactMethod(authenticator, splitted(1), authUrl, heads, body).asInstanceOf[Boolean]
-      if (result) {
-        logTrace(s"polling authentication of party=${authPartyID} passed")
+      val pollingAuthentication = new PollingAuthentication
+      val authResult = pollingAuthentication.authenticate(req)
+      if (authResult) {
+        logTrace(s"polling authentication of party=${req.getMetadata.getDst.getPartyId} passed")
       } else {
-        logError(s"polling authentication of party=${authPartyID} failed, please check polling client authentication info")
-        throw new IllegalArgumentException(s"polling authentication of party=${authPartyID} failed, please check polling client authentication info")
+        val errorInfo = new AuthenticationException(s"polling authentication of party=${req.getMetadata.getDst.getPartyId} failed, " +
+          s"please check polling client authentication info=${req.getMetadata.getTask.getModel.getDataKey}")
+        logError(s"polling authentication of party=${req.getMetadata.getDst.getPartyId} failed, please check polling client authentication info")
+        onError(errorInfo)
+        throw new AuthenticationException(s"polling authentication of party=${req.getMetadata.getDst.getPartyId} failed, please check polling client authentication info")
       }
     } else {
-      logDebug("polling authentication disable")
+      logDebug("polling authentication disabled")
     }
 
     pollingExchanger = new PollingExchanger()
